+++ examples/state/vsgtexturearray/vsgtexturearray.cpp
@@ -21,8 +21,8 @@ void updateBaseTexture(vsg::ubvec4Array2D& image, float value)
 
             float distance_from_center = vsg::length(delta);
 
-            float intensity = (sin(1.0 * angle + 30.0f * distance_from_center + 10.0 * value) + 1.0f) * 0.5f;
-            image.set(c, r, vsg::ubvec4(uint8_t(intensity * intensity * 255.0), uint8_t(intensity * 255.0), uint8_t(intensity * 255.0), 255));
+            float intensity = (sin(1.0f * angle + 30.0f * distance_from_center + 10.0f * value) + 1.0f) * 0.5f;
+            image.set(uint32_t(c), uint32_t(r), vsg::ubvec4(uint8_t(intensity * intensity * 255.0f), uint8_t(intensity * 255.0f), uint8_t(intensity * 255.0f), 255));
         }
     }
 }
@@ -42,8 +42,8 @@ void updateElevation(vsg::floatArray2D& heightField, float value)
 
             float distance_from_center = vsg::length(delta);
 
-            float intensity = (sin(1.0 * angle + 10.0f * distance_from_center + 2.0 * value) + 1.0f) * 0.5f;
-            heightField.set(c, r, intensity);
+            float intensity = (sin(1.0f * angle + 10.0f * distance_from_center + 2.0f * value) + 1.0f) * 0.5f;
+            heightField.set(uint32_t(c), uint32_t(r), intensity);
         }
     }
 }
@@ -57,7 +57,7 @@ vsg::ref_ptr<vsg::Node> createGeometry()
          {0.5f, 0.5f, 0.0f},
          {-0.5f, 0.5f, 0.0f}}); // VK_FORMAT_R32G32B32_SFLOAT, VK_VERTEX_INPUT_RATE_INSTANCE, VK_BUFFER_USAGE_VERTEX_BUFFER_BIT, VK_SHARING_MODE_EXCLUSIVE
 
-    auto colors = vsg::vec3Array::create(vertices->size(), vsg::vec3(1.0f, 1.0f, 1.0f));
+    auto colors = vsg::vec3Array::create(uint32_t(vertices->size()), vsg::vec3(1.0f, 1.0f, 1.0f));
 
     auto texcoords = vsg::vec2Array::create(
         {{0.0f, 0.0f},
